Fix bugs due to not restting atom tree model correctly, add comments.
This fixes a bug where the model would notify a layout change before
adding data, when it should be notifying a data change. *ResetModel
methods are now used to invalidate the model before we update the atoms,
which should fix issues with bad indices that could occur if an mp4 is
opened and nodes expanded, and then another file is loaded.

Also tweak some comments and asserts.